feat(api): remove all mock data and mock mode support
Deletes mock.ts and MockProvider.ts entirely. Extracts type definitions
to types.ts. All routes now call Coolify directly (502 on failure) and
DNS fails fast at startup if TECHNITIUM_URL/TOKEN are not configured.
