+++ api/src/routes/sites.ts
@@ -1,5 +1,5 @@
 import { Router, Request, Response } from 'express';
-import { SITES, getSite, addSite, removeSite, updateSite, DnsRecord } from '../data/mock';
+import { DnsRecord } from '../types';
 import {
   createCoolifyClient,
   CoolifyCreateApplicationPayload,
@@ -11,15 +11,10 @@ import { createDnsProvider, DnsOperationError } from '../services/dns';
 
 const router = Router();
 
-const useCoolifyMock = !process.env.COOLIFY_API_URL;
 const dnsProvider = createDnsProvider();
 
 // ── GET /api/sites — list all sites ──────────────────────────────────────────
 router.get('/', async (_req: Request, res: Response) => {
-  if (useCoolifyMock) {
-    res.status(200).json(SITES);
-    return;
-  }
   try {
     const client = createCoolifyClient()!;
     const applications = await client.listApplications();
@@ -31,24 +26,13 @@ router.get('/', async (_req: Request, res: Response) => {
     );
     res.status(200).json(sites);
   } catch (err) {
-    console.warn('[coolify] GET /applications failed, falling back to mock:', (err as Error).message);
-    res.setHeader('x-data-source', 'mock');
-    res.status(200).json(SITES);
+    console.error('[coolify] GET /applications failed:', (err as Error).message);
+    res.status(502).json({ error: 'Failed to retrieve sites from Coolify' });
   }
 });
 
 // ── GET /api/sites/:slug — single site detail ─────────────────────────────────
 router.get('/:slug', async (req: Request, res: Response) => {
-  if (useCoolifyMock) {
-    const site = getSite(req.params.slug);
-    if (!site) {
-      res.status(404).json({ error: 'Site not found' });
-      return;
-    }
-    const probe = site.domain ? await probeSite(site.domain).catch(() => null) : null;
-    res.status(200).json(probe ? { ...site, ...probe } : site);
-    return;
-  }
   try {
     const client = createCoolifyClient()!;
     const app = await client.getApplication(req.params.slug);
@@ -65,54 +49,14 @@ router.get('/:slug', async (req: Request, res: Response) => {
 
     res.status(200).json(mapSite(app, deployments, probe));
   } catch (err) {
-    console.warn(`[coolify] GET /applications/${req.params.slug} failed, falling back to mock:`, (err as Error).message);
-    const site = getSite(req.params.slug);
-    if (!site) {
-      res.status(404).json({ error: 'Site not found' });
-      return;
-    }
-    res.setHeader('x-data-source', 'mock');
-    res.status(200).json(site);
+    console.error(`[coolify] GET /applications/${req.params.slug} failed:`, (err as Error).message);
+    res.status(502).json({ error: 'Failed to retrieve site from Coolify' });
   }
 });
 
 // ── POST /api/sites — create a new site ──────────────────────────────────────
+// Requires full Coolify payload: name, git_repository, git_branch, server_uuid, destination_uuid
 router.post('/', async (req: Request, res: Response) => {
-  // Mock mode accepts the frontend's simplified shape: name, domain, gitRepo, gitBranch.
-  // Coolify mode requires the full CoolifyCreateApplicationPayload fields.
-  if (useCoolifyMock) {
-    const { name, domain, gitRepo, gitBranch } = req.body as {
-      name?: string;
-      domain?: string;
-      gitRepo?: string;
-      gitBranch?: string;
-    };
-    if (!name || !domain) {
-      res.status(400).json({ error: 'Missing required fields: name, domain' });
-      return;
-    }
-    const slug = name.trim().toLowerCase().replace(/\s+/g, '-').replace(/[^a-z0-9-]/g, '');
-    const now = new Date().toISOString();
-    const newSite = {
-      slug,
-      name: name.trim(),
-      domain: domain.trim(),
-      description: '',
-      repository: gitRepo?.trim() ?? '',
-      server: '',
-      overallStatus: 'pending' as const,
-      http: { reachable: false, statusCode: null, responseTimeMs: null, checkedAt: now },
-      ssl: { valid: false, expiresAt: null, daysUntilExpiry: null, issuer: null, checkedAt: now },
-      dns: { resolving: false, propagated: false, checkedAt: now },
-      dnsRecords: [],
-      deploys: [],
-      ...(gitBranch ? {} : {}),
-    };
-    addSite(newSite);
-    res.status(201).json(newSite);
-    return;
-  }
-
   const body = req.body as Partial<CoolifyCreateApplicationPayload>;
   if (!body.name || !body.git_repository || !body.git_branch || !body.server_uuid || !body.destination_uuid) {
     res.status(400).json({
@@ -135,64 +79,35 @@ router.post('/', async (req: Request, res: Response) => {
     const app = await client.createApplication(payload);
     res.status(201).json(mapSite(app, []));
   } catch (err) {
-    console.warn('[coolify] POST /applications failed:', (err as Error).message);
+    console.error('[coolify] POST /applications failed:', (err as Error).message);
     res.status(502).json({ error: 'Failed to create application via Coolify' });
   }
 });
 
 // ── DELETE /api/sites/:slug — delete a site ───────────────────────────────────
 router.delete('/:slug', async (req: Request, res: Response) => {
-  if (useCoolifyMock) {
-    const removed = removeSite(req.params.slug);
-    if (!removed) {
-      res.status(404).json({ error: 'Site not found' });
-      return;
-    }
-    res.status(204).send();
-    return;
-  }
   try {
     const client = createCoolifyClient()!;
     await client.deleteApplication(req.params.slug);
     res.status(204).send();
   } catch (err) {
-    console.warn(`[coolify] DELETE /applications/${req.params.slug} failed:`, (err as Error).message);
+    console.error(`[coolify] DELETE /applications/${req.params.slug} failed:`, (err as Error).message);
     res.status(502).json({ error: 'Failed to delete application via Coolify' });
   }
 });
 
 // ── PATCH /api/sites/:slug — update site settings ────────────────────────────
 //
-// Field alignment: the frontend Settings tab sends Site-typed fields
-// (repository, server, description) rather than Coolify API fields
-// (git_repository, build_pack, fqdn). Accepting the frontend's field names
-// here keeps the frontend decoupled from Coolify internals. In Coolify mode
-// we translate: repository → git_repository. The `server` field has no
-// Coolify equivalent and is only applied in mock mode.
+// Accepts both frontend Site fields (repository, description) and raw Coolify
+// fields (git_repository, build_pack, fqdn). repository → git_repository translation
+// keeps the frontend decoupled from Coolify internals.
 router.patch('/:slug', async (req: Request, res: Response) => {
-  // Accept both frontend Site fields and raw Coolify fields.
   const body = req.body as Partial<CoolifyUpdateApplicationPayload & {
     repository?: string;
     server?: string;
   }>;
   if (Object.keys(body).length === 0) {
     res.status(400).json({ error: 'Request body must include at least one field to update' });
-    return;
-  }
-  if (useCoolifyMock) {
-    const site = getSite(req.params.slug);
-    if (!site) {
-      res.status(404).json({ error: 'Site not found' });
-      return;
-    }
-    const updates: Partial<import('../data/mock').Site> = {};
-    if (body.repository !== undefined) updates.repository = body.repository;
-    if (body.git_repository !== undefined) updates.repository = body.git_repository;
-    if (body.server !== undefined) updates.server = body.server;
-    if (body.description !== undefined) updates.description = body.description;
-    if (body.name !== undefined) updates.name = body.name;
-    const updated = updateSite(req.params.slug, updates);
-    res.status(200).json(updated);
     return;
   }
   try {
@@ -209,35 +124,54 @@ router.patch('/:slug', async (req: Request, res: Response) => {
     const app = await client.updateApplication(req.params.slug, payload);
     res.status(200).json(mapSite(app, []));
   } catch (err) {
-    console.warn(`[coolify] PATCH /applications/${req.params.slug} failed:`, (err as Error).message);
+    console.error(`[coolify] PATCH /applications/${req.params.slug} failed:`, (err as Error).message);
     res.status(502).json({ error: 'Failed to update application via Coolify' });
   }
 });
 
 // ── GET /api/sites/:slug/dns — DNS records for a site ────────────────────────
 router.get('/:slug/dns', async (req: Request, res: Response) => {
-  const site = getSite(req.params.slug);
-  if (!site) { res.status(404).json({ error: 'Site not found' }); return; }
   try {
-    const records = await dnsProvider.getRecords(site.domain);
+    const client = createCoolifyClient()!;
+    const app = await client.getApplication(req.params.slug);
+    const domain = app.fqdn
+      ? app.fqdn.split(',')[0].trim().replace(/^https?:\/\//, '')
+      : '';
+    if (!domain) {
+      res.status(422).json({ error: 'Site has no domain configured' });
+      return;
+    }
+    const records = await dnsProvider.getRecords(domain);
     res.status(200).json(records);
   } catch (err) {
-    console.warn('[dns] getRecords failed:', err instanceof DnsOperationError ? err.originalCause : err);
-    res.status(500).json({ error: 'DNS operation failed' });
+    if (err instanceof DnsOperationError) {
+      console.warn('[dns] getRecords failed:', err.originalCause);
+      res.status(500).json({ error: 'DNS operation failed' });
+    } else {
+      console.error(`[coolify] GET /applications/${req.params.slug} for DNS failed:`, (err as Error).message);
+      res.status(502).json({ error: 'Failed to retrieve site from Coolify' });
+    }
   }
 });
 
 // ── POST /api/sites/:slug/dns — add a DNS record ─────────────────────────────
 router.post('/:slug/dns', async (req: Request, res: Response) => {
-  const site = getSite(req.params.slug);
-  if (!site) { res.status(404).json({ error: 'Site not found' }); return; }
-  const body = req.body as Partial<DnsRecord>;
-  if (!body.type || !body.name || !body.value || !body.ttl) {
-    res.status(400).json({ error: 'Missing required fields: type, name, value, ttl' });
-    return;
-  }
   try {
-    const record = await dnsProvider.addRecord(site.domain, {
+    const client = createCoolifyClient()!;
+    const app = await client.getApplication(req.params.slug);
+    const domain = app.fqdn
+      ? app.fqdn.split(',')[0].trim().replace(/^https?:\/\//, '')
+      : '';
+    if (!domain) {
+      res.status(422).json({ error: 'Site has no domain configured' });
+      return;
+    }
+    const body = req.body as Partial<DnsRecord>;
+    if (!body.type || !body.name || !body.value || !body.ttl) {
+      res.status(400).json({ error: 'Missing required fields: type, name, value, ttl' });
+      return;
+    }
+    const record = await dnsProvider.addRecord(domain, {
       type: body.type,
       name: body.name,
       value: body.value,
@@ -246,49 +180,69 @@ router.post('/:slug/dns', async (req: Request, res: Response) => {
     });
     res.status(201).json(record);
   } catch (err) {
-    console.warn('[dns] addRecord failed:', err instanceof DnsOperationError ? err.originalCause : err);
-    res.status(500).json({ error: 'DNS operation failed' });
+    if (err instanceof DnsOperationError) {
+      console.warn('[dns] addRecord failed:', err.originalCause);
+      res.status(500).json({ error: 'DNS operation failed' });
+    } else {
+      console.error(`[coolify] GET /applications/${req.params.slug} for DNS failed:`, (err as Error).message);
+      res.status(502).json({ error: 'Failed to retrieve site from Coolify' });
+    }
   }
 });
 
 // ── PUT /api/sites/:slug/dns/:id — update a DNS record ───────────────────────
 router.put('/:slug/dns/:id', async (req: Request, res: Response) => {
-  const site = getSite(req.params.slug);
-  if (!site) { res.status(404).json({ error: 'Site not found' }); return; }
   try {
+    const client = createCoolifyClient()!;
+    const app = await client.getApplication(req.params.slug);
+    const domain = app.fqdn
+      ? app.fqdn.split(',')[0].trim().replace(/^https?:\/\//, '')
+      : '';
+    if (!domain) {
+      res.status(422).json({ error: 'Site has no domain configured' });
+      return;
+    }
     const updates = req.body as Partial<Omit<DnsRecord, 'id'>>;
-    const record = await dnsProvider.updateRecord(site.domain, req.params.id, updates);
+    const record = await dnsProvider.updateRecord(domain, req.params.id, updates);
     res.status(200).json(record);
   } catch (err) {
-    console.warn('[dns] updateRecord failed:', err instanceof DnsOperationError ? err.originalCause : err);
-    res.status(500).json({ error: 'DNS operation failed' });
+    if (err instanceof DnsOperationError) {
+      console.warn('[dns] updateRecord failed:', err.originalCause);
+      res.status(500).json({ error: 'DNS operation failed' });
+    } else {
+      console.error(`[coolify] GET /applications/${req.params.slug} for DNS failed:`, (err as Error).message);
+      res.status(502).json({ error: 'Failed to retrieve site from Coolify' });
+    }
   }
 });
 
 // ── DELETE /api/sites/:slug/dns/:id — delete a DNS record ────────────────────
 router.delete('/:slug/dns/:id', async (req: Request, res: Response) => {
-  const site = getSite(req.params.slug);
-  if (!site) { res.status(404).json({ error: 'Site not found' }); return; }
   try {
-    await dnsProvider.deleteRecord(site.domain, req.params.id);
+    const client = createCoolifyClient()!;
+    const app = await client.getApplication(req.params.slug);
+    const domain = app.fqdn
+      ? app.fqdn.split(',')[0].trim().replace(/^https?:\/\//, '')
+      : '';
+    if (!domain) {
+      res.status(422).json({ error: 'Site has no domain configured' });
+      return;
+    }
+    await dnsProvider.deleteRecord(domain, req.params.id);
     res.status(204).send();
   } catch (err) {
-    console.warn('[dns] deleteRecord failed:', err instanceof DnsOperationError ? err.originalCause : err);
-    res.status(500).json({ error: 'DNS operation failed' });
+    if (err instanceof DnsOperationError) {
+      console.warn('[dns] deleteRecord failed:', err.originalCause);
+      res.status(500).json({ error: 'DNS operation failed' });
+    } else {
+      console.error(`[coolify] GET /applications/${req.params.slug} for DNS failed:`, (err as Error).message);
+      res.status(502).json({ error: 'Failed to retrieve site from Coolify' });
+    }
   }
 });
 
 // ── GET /api/sites/:slug/deployments — deployment history ────────────────────
 router.get('/:slug/deployments', async (req: Request, res: Response) => {
-  if (useCoolifyMock) {
-    const site = getSite(req.params.slug);
-    if (!site) {
-      res.status(404).json({ error: 'Site not found' });
-      return;
-    }
-    res.status(200).json(site.deploys);
-    return;
-  }
   try {
     const client = createCoolifyClient()!;
     const app = await client.getApplication(req.params.slug);
@@ -296,73 +250,33 @@ router.get('/:slug/deployments', async (req: Request, res: Response) => {
     const deploys = deployments.map((d) => mapDeploy(d, app.git_branch));
     res.status(200).json(deploys);
   } catch (err) {
-    console.warn(`[coolify] GET deployments for ${req.params.slug} failed, falling back to mock:`, (err as Error).message);
-    const site = getSite(req.params.slug);
-    if (!site) {
-      res.status(404).json({ error: 'Site not found' });
-      return;
-    }
-    res.setHeader('x-data-source', 'mock');
-    res.status(200).json(site.deploys);
+    console.error(`[coolify] GET deployments for ${req.params.slug} failed:`, (err as Error).message);
+    res.status(502).json({ error: 'Failed to retrieve deployments from Coolify' });
   }
 });
 
 // ── POST /api/sites/:slug/deploy — trigger a deploy ──────────────────────────
 router.post('/:slug/deploy', async (req: Request, res: Response) => {
-  if (useCoolifyMock) {
-    const site = getSite(req.params.slug);
-    if (!site) {
-      res.status(404).json({ error: 'Site not found' });
-      return;
-    }
-    res.status(202).json({ jobId: `job-${Date.now()}` });
-    return;
-  }
   try {
     const client = createCoolifyClient()!;
     const result = await client.triggerDeploy(req.params.slug);
     res.status(202).json({ jobId: result.deployment_uuid, status: result.status, message: result.message });
   } catch (err) {
-    console.warn(`[coolify] POST /deploy for ${req.params.slug} failed:`, (err as Error).message);
+    console.error(`[coolify] POST /deploy for ${req.params.slug} failed:`, (err as Error).message);
     res.status(502).json({ error: 'Failed to trigger deploy via Coolify' });
   }
 });
 
 // ── GET /api/sites/:slug/deployments/:id/log — deploy log lines ───────────────
 router.get('/:slug/deployments/:id/log', async (req: Request, res: Response) => {
-  if (useCoolifyMock) {
-    const site = getSite(req.params.slug);
-    if (!site) {
-      res.status(404).json({ error: 'Site not found' });
-      return;
-    }
-    const deploy = site.deploys.find((d) => d.id === req.params.id);
-    if (!deploy) {
-      res.status(404).json({ error: 'Deployment not found' });
-      return;
-    }
-    res.status(200).json(deploy.logLines);
-    return;
-  }
   try {
     const client = createCoolifyClient()!;
     const deployment = await client.getDeployment(req.params.id);
     const mapped = mapDeploy(deployment, '');
     res.status(200).json(mapped.logLines);
   } catch (err) {
-    console.warn(`[coolify] GET deployment log for ${req.params.id} failed, falling back to mock:`, (err as Error).message);
-    const site = getSite(req.params.slug);
-    if (!site) {
-      res.status(404).json({ error: 'Site not found' });
-      return;
-    }
-    const deploy = site.deploys.find((d) => d.id === req.params.id);
-    if (!deploy) {
-      res.status(404).json({ error: 'Deployment not found' });
-      return;
-    }
-    res.setHeader('x-data-source', 'mock');
-    res.status(200).json(deploy.logLines);
+    console.error(`[coolify] GET deployment log for ${req.params.id} failed:`, (err as Error).message);
+    res.status(502).json({ error: 'Failed to retrieve deployment log from Coolify' });
   }
 });
